supplicant: don't allow scans to trigger a supplicant roaming decision
Just because the user requested a scan doesn't mean the supplicant should
use the result of that scan to jump to an AP that's slightly better than
the current one.  Let the supplicant handle when it's supposed to roam
based on it's own logic, not random scans from users or NM clients.