platform: preserve order of objects during dump
NMPCache can preserve the order of the objects. Until now, we
however didn't care.

During a dump, force appending new objects at the end. That ensures,
correct ordering during the dump.

Note, that we don't really care about it, except (soon) for routes.

Also, since we track objects in distinct indexes that partition the set
of all objects, it is much less clear what the order for those indexes
really means outside the initial dump. Meaning: when we get events that
add new objects, it is unclear where to insert them.